Разделяет js-код на модули
